requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
beautifulsoup4==4.11.1
certifi==2022.6.15.2; python_full_version >= '3.6.0'
charset-normalizer==2.1.1; python_full_version >= '3.6.0'
idna==3.4; python_version >= '3.5'
psycopg2-binary==2.9.3
pyaes==1.6.1
pyrogram==2.0.51
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1; python_full_version >= '3.6.0'
tgcrypto==1.2.3
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
